pyproject.toml
[tool.poetry]
name = "toy-system-resource-monitor"
version = "0.1.0"
description = ""
authors = ["lyg6772 <lyg6772@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ujson = "^5.10.0"
kafka-python-ng = "^2.2.2"
python-dotenv = "^1.0.1"
psutil = "^6.0.0"
gputil = "^1.4.0"
schedule = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
